requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
chromadb>=0.4.18
transformers>=4.36.0
anthropic>=0.7.0
python-multipart>=0.0.6
openai>=1.0.0
